Module aws_sdk_grafana::error

Errors that can occur when calling the service.

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You do not have sufficient permissions to perform this action.

AssociateLicenseError

Error type for the AssociateLicense operation.

ConflictException

A resource was in an inconsistent state during an update or a deletion.

CreateWorkspaceError

Error type for the CreateWorkspace operation.

DeleteWorkspaceError

Error type for the DeleteWorkspace operation.

DescribeWorkspaceAuthenticationError

Error type for the DescribeWorkspaceAuthentication operation.

DescribeWorkspaceError

Error type for the DescribeWorkspace operation.

DisassociateLicenseError

Error type for the DisassociateLicense operation.

InternalServerException

Unexpected error while processing the request. Retry the request.

ListPermissionsError

Error type for the ListPermissions operation.

ListWorkspacesError

Error type for the ListWorkspaces operation.

ResourceNotFoundException

The request references a resource that does not exist.

ServiceQuotaExceededException

The request would cause a service quota to be exceeded.

ThrottlingException

The request was denied because of request throttling. Retry the request.

UpdatePermissionsError

Error type for the UpdatePermissions operation.

UpdateWorkspaceAuthenticationError

Error type for the UpdateWorkspaceAuthentication operation.

UpdateWorkspaceError

Error type for the UpdateWorkspace operation.

ValidationException

The value of a parameter in the request caused an error.

Enums

AssociateLicenseErrorKind

Types of errors that can occur for the AssociateLicense operation.

CreateWorkspaceErrorKind

Types of errors that can occur for the CreateWorkspace operation.

DeleteWorkspaceErrorKind

Types of errors that can occur for the DeleteWorkspace operation.

DescribeWorkspaceAuthenticationErrorKind

Types of errors that can occur for the DescribeWorkspaceAuthentication operation.

DescribeWorkspaceErrorKind

Types of errors that can occur for the DescribeWorkspace operation.

DisassociateLicenseErrorKind

Types of errors that can occur for the DisassociateLicense operation.

ListPermissionsErrorKind

Types of errors that can occur for the ListPermissions operation.

ListWorkspacesErrorKind

Types of errors that can occur for the ListWorkspaces operation.

UpdatePermissionsErrorKind

Types of errors that can occur for the UpdatePermissions operation.

UpdateWorkspaceAuthenticationErrorKind

Types of errors that can occur for the UpdateWorkspaceAuthentication operation.

UpdateWorkspaceErrorKind

Types of errors that can occur for the UpdateWorkspace operation.