Module aws_sdk_grafana::model

Data structures used by operation inputs/outputs.

Modules

assertion_attributes

See AssertionAttributes

authentication_description

See AuthenticationDescription

authentication_summary

See AuthenticationSummary

aws_sso_authentication

See AwsSsoAuthentication

permission_entry

See PermissionEntry

role_values

See RoleValues

saml_authentication

See SamlAuthentication

saml_configuration

See SamlConfiguration

update_error

See UpdateError

update_instruction

See UpdateInstruction

user

See User

validation_exception_field

See ValidationExceptionField

workspace_description

See WorkspaceDescription

workspace_summary

See WorkspaceSummary

Structs

AssertionAttributes

A structure that defines which attributes in the IdP assertion are to be used to define

AuthenticationDescription

A structure containing information about the user authentication methods used by the workspace.

AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods

AwsSsoAuthentication

A structure containing information about how this workspace works with

PermissionEntry

A structure containing the identity of one user or group and the Admin

RoleValues

This structure defines which groups defined in the SAML assertion attribute are to be mapped

SamlAuthentication

A structure containing information about how this workspace works with

SamlConfiguration

A structure containing information about how this workspace works with

UpdateError

A structure containing information about one error encountered while performing an

UpdateInstruction

Contains the instructions for one Grafana role permission update in a

User

A structure that specifies one user or group in the workspace.

ValidationExceptionField

A structure that contains information about a request parameter that caused an error.

WorkspaceDescription

A structure containing information about an Amazon Managed Grafana workspace in your account.

WorkspaceSummary

A structure that contains some information about one workspace in the account.

Enums

AccountAccessType

AuthenticationProviderTypes

DataSourceType

IdpMetadata

A structure containing the identity provider (IdP) metadata used to integrate the

LicenseType

NotificationDestinationType

PermissionType

Role

SamlConfigurationStatus

UpdateAction

UserType

ValidationExceptionReason

WorkspaceStatus